Chore: add tests and refactor files
