+++ references/cuegen/generator.go
@@ -17,8 +17,11 @@ limitations under the License.
 package cuegen
 
 import (
+	"fmt"
 	goast "go/ast"
+	gotypes "go/types"
 	"io"
+	"strings"
 
 	cueast "cuelang.org/go/cue/ast"
 	"cuelang.org/go/cue/ast/astutil"
@@ -98,4 +101,52 @@ func (g *Generator) write(w io.Writer, f *cueast.File) error {
 
 	_, err = w.Write(b)
 	return err
+}
+
+// loadPackage loads a package from given path.
+func loadPackage(p string) (*packages.Package, error) {
+	cfg := &packages.Config{
+		Mode: packages.NeedName | packages.NeedFiles | packages.NeedCompiledGoFiles |
+			packages.NeedImports | packages.NeedTypes | packages.NeedTypesSizes |
+			packages.NeedSyntax | packages.NeedTypesInfo | packages.NeedDeps |
+			packages.NeedModule,
+	}
+
+	pkgs, err := packages.Load(cfg, []string{p}...)
+	if err != nil {
+		return nil, err
+	}
+	if len(pkgs) != 1 {
+		return nil, fmt.Errorf("expected one package, got %d", len(pkgs))
+	}
+
+	// only need to check the first package
+	pkg := pkgs[0]
+	if pkg.Errors != nil {
+		errs := make([]string, 0, len(pkg.Errors))
+		for _, e := range pkg.Errors {
+			errs = append(errs, fmt.Sprintf("\t%s: %v", pkg.PkgPath, e))
+		}
+		return nil, fmt.Errorf("could not load Go packages:\n%s", strings.Join(errs, "\n"))
+	}
+
+	return pkg, nil
+}
+
+type typeInfo map[gotypes.Type]*goast.StructType
+
+func getTypeInfo(p *packages.Package) typeInfo {
+	m := make(typeInfo)
+
+	for _, f := range p.Syntax {
+		goast.Inspect(f, func(n goast.Node) bool {
+			// record all struct types
+			if t, ok := n.(*goast.StructType); ok {
+				m[p.TypesInfo.TypeOf(t)] = t
+			}
+			return true
+		})
+	}
+
+	return m
 }
